DefiIn/requirements.txt
web3==6.15.0            
aiohttp==3.9.3           
python-dotenv==1.0.0    
eth-account==0.11.0      
pytest==8.0.2
pytest-asyncio==0.23.5
requests-mock==1.11.0    
eth-abi==4.2.1           
tqdm==4.66.1            
python-dateutil==2.8.2   